Custom hook for MovieScreen
